Merge branch 'master' into VP_CONFIGURABLE_FIELDS

+++ WEB-INF/src/com/krishagni/catissueplus/core/biospecimen/repository/SpecimenListDao.java
@@ -22,7 +22,7 @@ public interface SpecimenListDao extends Dao<SpecimenList> {
 
 	public SpecimenList getDefaultSpecimenList(Long userId);
 	
-	public Long getListSpecimensCount(Long listId);
+	public int getListSpecimensCount(Long listId);
 	
 	public void deleteSpecimenList(SpecimenList listId);
 }
